Extract repeated diag func for k8s to a utility type.
